api/requirements.txt
# API-specific requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Caching
redis==5.0.1

# ML
numpy==1.26.3
pandas==2.2.0
lightgbm==4.3.0
scikit-learn==1.4.0
joblib==1.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
